sideproject/oh-manager-scheduler/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.3
dependency-injector==4.41.0
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
inflection==0.5.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langsmith==0.0.84
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.10.0
openapi==1.1.0
packaging==23.2
pydantic==2.6.0
pydantic_core==2.16.1
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
slack_sdk==3.26.2
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.5
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
yarl==1.9.4